+++ doc/Changelog.md
@@ -1,8 +1,10 @@
 Changelog {#changelog}
 =========
 
-# Release 0.3 (git-master/current) (10-Jul-2015)
+# git master {#master}
 
+* [14496](https://bbpcode.epfl.ch/code/#/c/14496/)
+  New synapse loader to visualize synapse densities
 * [14181](https://bbpcode.epfl.ch/code/#/c/14181/)
   New dye curve parameter for VSD reports in the Livre data source
 * [14203](https://bbpcode.epfl.ch/code/#/c/14203/)